Added management of data size in frame to avoid filling the end of the file with garbage data.
Corrected some bugs in error injection.
Added method description.

+++ TransmissionSimulation/TransmissionSimulation/MainForm.cs
@@ -64,7 +64,7 @@ namespace TransmissionSimulation
             {
                 if (errorPosition.Validate() && errorPosition.Value != -1)
                 {
-                    cable.IndicesInversions.Add(errorPosition.Value);
+                    cable.IndicesInversions.Add((int)errorPosition.Value);
                 }
             }
         }
